+++ demisto_sdk/commands/validate/validators/DO_validators/DO101_docker_image_is_not_demisto.py
@@ -12,10 +12,10 @@ from demisto_sdk.commands.validate.validators.base_validator import (
 ContentTypes = Union[Integration, Script]
 
 
-class DockerImageIsNotNativeImageValidator(BaseValidator[ContentTypes]):
-    error_code = "DO110"
-    description = "Validate that the given content item uses a docker image that is not the native image."
-    error_message = "The docker image {0} is native-image, do not use a native-image"
+class DockerImageIsNotDemistoValidator(BaseValidator[ContentTypes]):
+    error_code = "DO101"
+    description = "Validate that the given content-item uses demisto docker image"
+    error_message = "docker image {0} is not a valid demisto-image, demisto image should start with demisto/"
     related_field = "Docker image"
     is_auto_fixable = False
 
@@ -28,5 +28,5 @@ class DockerImageIsNotNativeImageValidator(BaseValidator[ContentTypes]):
             )
             for content_item in content_items
             if not content_item.is_javascript
-            and content_item.docker_image_object.is_native_image
+            and not content_item.docker_image_object.is_demisto_repository
         ]
