+++ demisto_sdk/commands/validate/validators/DO_validators/DO104_docker_image_does_not_exist_in_yml.py
